% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walley1996.R
\name{dbetabinom}
\alias{dbetabinom}
\alias{pbetabinom}
\title{Beta-Binomial Distribution}
\usage{
dbetabinom(i, M, x, s, N, tA)

pbetabinom(M, x, s, N, y)
}
\arguments{
\item{i}{the number of occurrences of A in M future trials}

\item{M}{a number of future trials}

\item{x}{the number of occurrence of A in N previous trials}

\item{s}{learning parameter}

\item{N}{a total number of previous trials}

\item{tA}{the prior probability of A under the Dirichlet prior}

\item{y}{the number of occurrences of A in M future trials}
}
\description{
Density function of beta-binomial distribution
}
