% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walley1996.R
\name{dbetadif}
\alias{dbetadif}
\title{Distribution of Difference of Two Proportions}
\usage{
dbetadif(x, a1, b1, a2, b2)
}
\arguments{
\item{x}{difference of two beta distributions}

\item{a1}{shape 1 parameter of Beta distribution with control}

\item{b1}{shape 2 parameter of Beta distribution with control}

\item{a2}{shape 1 parameter of Beta distribution with treatment}

\item{b2}{shape 2 parameter of Beta distribution with treatment}
}
\value{
\code{betadif} gives a scalar value of density.
}
\description{
Distribution of Difference of Two Proportions
}
\references{
Chen, Y., & Luo, S. (2011). A few remarks on 'Statistical distribution of the difference of two proportions' by Nadarajah and Kotz, Statistics in Medicine 2007; 26 (18): 3518-3523. Statistics in Medicine, 30(15), 1913-1915.
}
