% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walley1996.R
\name{ibm}
\alias{ibm}
\title{impreicse beta model}
\usage{
ibm(n = 10, m = 6, s0 = 2, xlab1, main1)
}
\arguments{
\item{n}{total of observations}

\item{m}{number of observations seen}

\item{s0}{learning parameter}

\item{xlab1}{label}

\item{main1}{main}
}
\description{
imprecise beta model
}
\references{
Walley, P. (1996), Inferences from Multinomial Data: Learning About a Bag of Marbles. Journal of the Royal Statistical Society: Series B (Methodological), 58: 3-34. https://doi.org/10.1111/j.2517-6161.1996.tb02065.x
}
