% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walley1996.R
\name{idm}
\alias{idm}
\alias{hpd}
\title{imprecise Dirichlet model}
\usage{
idm(nj, s = 1, N, tj = NA_real_, k, cA = 1)

hpd(
  alpha = 3,
  beta = 5,
  p = 0.95,
  tolerance = 1e-04,
  maxiter = 100,
  verbose = FALSE
)
}
\arguments{
\item{nj}{number of observations in the j-th category}

\item{s}{learning parameter}

\item{N}{a total number of drawings}

\item{tj}{proability}

\item{k}{number of elements in Omega}

\item{cA}{the number of elements in A}

\item{alpha}{shape1 parameter of beta distribution}

\item{beta}{shape2 parameter of beta distribution}

\item{p}{credible level}

\item{tolerance}{tolerance level}

\item{maxiter}{maximum number of iterations}

\item{verbose}{silence}
}
\description{
lower and upper posterior probabilities

Find the highest posterior density interval
}
\examples{
idm(nj=1, N=6, s=2, k=4)
idm(nj=1, N=6, s=1, k=4)
x <- hpd(alpha=3, beta=5, p=0.95) # c(0.0031, 0.6587) when s=2
# round(x,4); x*(1-x)^5
}
\references{
Walley, P. (1996), Inferences from Multinomial Data: Learning About a Bag of Marbles. Journal of the Royal Statistical Society: Series B (Methodological), 58: 3-34. https://doi.org/10.1111/j.2517-6161.1996.tb02065.x
}
