% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walley1996.R
\name{idm}
\alias{idm}
\alias{hpd}
\title{imprecise Dirichlet model}
\usage{
idm(nj, s = 1, N, tj = NA)

hpd(alpha = 3, beta = 5, p = 0.95, tolerance = 1e-04, maxiter = 100)
}
\arguments{
\item{nj}{number of observations in the j-th category}

\item{s}{learning parameter}

\item{N}{a total number of drawings}

\item{tj}{proability}

\item{alpha}{shape1 parameter of beta distribution}

\item{beta}{shape2 parameter of beta distribution}

\item{p}{credible level}

\item{tolerance}{tolerance level}

\item{maxiter}{maximum number of iterations}
}
\description{
lower and upper posterior probabilities

Find the highest posterior density interval
}
\examples{
idm(nj=1, N=6, s=2)
idm(nj=1, N=6, s=1)
}
